requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
pandas==2.1.4
timm==1.0.7
opencv-python==4.9.0.80
onnxruntime==1.18.0
streamlit==1.31.0
Pillow
opencv-python-headless
tqdm
